Implement durable account banking and repair episode navigation
Share the bank across characters on one login with atomic inventory, wallet,
registry and journal commits, replay-safe exchanges and a native Windows panel.
Wait for durable bank and reform acknowledgements and reject partial SQL reads.

Restore episode doorways and re-entry, fix navigation distances and approaches,
add missing quest guides, and repair client archive priority, resources and
preflight validation. Include migration instructions and audit reports.

Validation: 26 source release checks on the exported staged tree, native
navigation tests with ASan/UBSan, and eight Windows launcher checks. SQL crash
recovery, real server sessions and native client evidence are recorded in the
bank and quest reports; complete rendered gameplay remains unverified.

+++ src/map/navi.hpp
@@ -36,7 +36,7 @@ struct navi_link {
 #define MAX_WALKPATH_NAVI 1024
 
 struct navi_walkpath_data {
-	uint8 path_len, path_pos;
+	uint16 path_len, path_pos;
 	uint8 path[MAX_WALKPATH_NAVI];
 };
